refactor: move settings to services
