+++ packages/photo-companion/src/lib/context/city-lights.tsx
@@ -6,8 +6,7 @@ import type { Accessor } from "solid-js";
 import type { LightsCity } from "types";
 
 import { useDatetime } from "~/lib/hooks";
-
-import { useSettings } from "./settings";
+import { useSettings } from "~/services/settings";
 
 interface CityLightsContextType extends Omit<LightsProvider, "city" | "year"> {
 	getCity: Accessor<LightsCity>;
