+++ packages/photo-companion/src/lib/context/translation/translation.context.tsx
@@ -13,7 +13,7 @@ import {
 } from "utils/formatters";
 import { isNullable } from "utils/validators";
 
-import { useSettings } from "../settings";
+import { useSettings } from "~/services/settings";
 
 import type { TranslationContextType } from "./translation.types";
