+++ packages/photo-companion/src/lib/hooks/use-supports.ts
@@ -1,6 +1,6 @@
 import { isSupportedCity } from "lights-schedule";
 
-import { useSettings } from "~/lib/context/settings";
+import { useSettings } from "~/services/settings";
 
 export function useSupportsLights() {
 	const { settings } = useSettings();
