+++ packages/photo-companion/src/routes/lights/components/lights-timeline.tsx
@@ -5,8 +5,8 @@ import type { EventName } from "types";
 import { Timeline, TimelineEvent } from "~/lib/components";
 import { ROUTES } from "~/lib/consts";
 import { useCityLights } from "~/lib/context/city-lights";
-import { useSettings } from "~/lib/context/settings";
 import { useDatetime, useTimelineProvider } from "~/lib/hooks";
+import { useSettings } from "~/services/settings";
 import { getSunEvents } from "~/services/sun";
 
 const TIMELINE_EVENT_SET = new Set<EventName>([
