+++ packages/photo-companion/src/routes/settings/settings.const.ts
@@ -3,8 +3,8 @@ import { SUPPORTED_BRIDGES_NAME_SET } from "bridge-schedule";
 import type { City } from "types";
 import type { InputSelectOption } from "ui";
 
-import type { SettingsStore } from "~/lib/context/settings";
 import type { Translation } from "~/lib/context/translation";
+import type { SettingsStore } from "~/services/settings";
 
 export const cityOptions: City[] = [
 	"MOSCOW",
