+++ packages/photo-companion/src/routes/settings/settings.context.tsx
@@ -4,8 +4,8 @@ import { isNullable } from "utils/validators";
 
 import type { City } from "types";
 
-import { SETTINGS_DEFAULT, type SettingsStore, useSettings } from "~/lib/context/settings";
-import { SETTINGS_CITY_PRESETS } from "~/lib/context/settings/settings.const";
+import { SETTINGS_DEFAULT, type SettingsStore, useSettings } from "~/services/settings";
+import { SETTINGS_CITY_PRESETS } from "~/services/settings/settings.const";
 
 import { FORM_INPUT_NAME } from "./settings.const";
