+++ packages/photo-companion/src/routes/settings/settings.form.tsx
@@ -1,7 +1,7 @@
 import { Button, Fieldset, Form, InputRadio } from "ui";
 
-import { SETTINGS_CITY_PRESETS } from "~/lib/context/settings";
 import { useTranslation } from "~/lib/context/translation";
+import { SETTINGS_CITY_PRESETS } from "~/services/settings";
 
 import { InputEventFilters, InputGeolocation, InputTabsSelect } from "./components";
 import { CITY_OPTIONS, FORM_INPUT_NAME, LANGUAGE_OPTIONS } from "./settings.const";
@@ -11,7 +11,7 @@ import styles from "./settings.module.css";
 
 export function SettingsForm() {
 	const { t } = useTranslation();
-	
+
 	const { handleFormChange, handleReset, handleSubmit, settingsStore } = useSettingsPage();
 
 	return (
