+++ packages/photo-companion/src/routes/sun/sun-timeline.tsx
@@ -2,8 +2,8 @@ import { For } from "solid-js";
 
 import { Timeline, TimelineEvent } from "~/lib/components";
 import { ROUTES } from "~/lib/consts";
-import { useSettings } from "~/lib/context/settings";
 import { useTimelineProvider } from "~/lib/hooks";
+import { useSettings } from "~/services/settings";
 import { getSunEvents } from "~/services/sun";
 
 interface SunTimelineProps {
