+++ packages/photo-companion/src/routes/timeline/hooks/use-disabled-timeline.ts
@@ -1,4 +1,4 @@
-import { useSettings } from "~/lib/context/settings";
+import { useSettings } from "~/services/settings";
 
 export function useDisabledTimeline() {
 	const { settings } = useSettings();
