+++ packages/photo-companion/src/routes/timeline/hooks/use-timeline-events.ts
@@ -2,10 +2,10 @@ import { getBridgeEvents } from "bridge-schedule";
 import { isNullable } from "utils/validators";
 
 import { useCityLights } from "~/lib/context/city-lights";
-import { useSettings } from "~/lib/context/settings";
 import { useSupportsBridges, useSupportsLights } from "~/lib/hooks";
 import { useDatetime, useTimelineProvider } from "~/lib/hooks";
 import { getMoonEvents } from "~/services/moon";
+import { useSettings } from "~/services/settings";
 import { getSunEvents } from "~/services/sun";
 
 import { useTimelineFilters } from "./use-timeline-filters";
