+++ packages/photo-companion/src/routes/timeline/hooks/use-timeline-filters.ts
@@ -3,7 +3,7 @@ import { isNullable } from "utils/validators";
 
 import type { EventName } from "types";
 
-import { useSettings } from "~/lib/context/settings";
+import { useSettings } from "~/services/settings";
 
 /**
  * Returns a Set of user blocked timeline events.
