+++ packages/photo-companion/src/routes/timeline/timeline.page.tsx
@@ -6,9 +6,9 @@ import type { EventGroupName } from "types";
 
 import { Timeline, TimelineEvent, TimelineEventEmpty, TimelineGroup } from "~/lib/components";
 import { CityLightsProvider } from "~/lib/context/city-lights";
-import { useSettings } from "~/lib/context/settings";
 import { useTranslation } from "~/lib/context/translation";
 import { useDatetime } from "~/lib/hooks";
+import { useSettings } from "~/services/settings";
 
 import { useDisabledTimeline, useTimelineEvents } from "./hooks";
