+++ packages/photo-companion/src/services/navigation/navigation.state.ts
@@ -2,8 +2,8 @@ import { useLocation, useNavigate } from "@solidjs/router";
 import { createMemo } from "solid-js";
 
 import { ROUTES } from "~/lib/consts";
-import { useSettings } from "~/lib/context/settings";
 import { useSupportsBridges, useSupportsLights } from "~/lib/hooks";
+import { useSettings } from "~/services/settings";
 
 import { NAVIGATION_TAB_DATA } from "./navigation-service.consts";
