+++ packages/photo-companion/src/services/settings/settings.const.ts
