+++ packages/photo-companion/src/services/settings/settings.types.ts
